sortBy option in listRequest
