+++ src/components/Switch.tsx
@@ -1,10 +1,10 @@
 import React from "react";
 import { useSelector } from "react-redux";
 import Main from "./Main/Main";
-import Patient from "./Patient/Patient";
+import PatientList from "./Patient/PatientList";
 import Appointment from "./Appointment/Appointment";
 import Treatment from "./Treatment/Treatment";
-import EditPatient from "./Patient/EditPatient";
+import PatientEdit from "./Patient/PatientEdit";
 import Scanner from "./Scanner/Scanner";
 import { getScreen } from "../redux/reducers";
 
@@ -12,11 +12,11 @@ const Switch: React.FunctionComponent = () => {
   const screen = useSelector(getScreen);
   switch (screen) {
     case "PATIENT":
-      return <Patient />;
+      return <PatientList />;
     case "ADD_PATIENT":
-      return <EditPatient />;
+      return <PatientEdit />;
     case "EDIT_PATIENT":
-      return <EditPatient />;
+      return <PatientEdit />;
     case "APPOINTMENT":
       return <Appointment />;
     case "TREATMENT":
